# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoanLedger API - Loan management with interest and commission tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user basic info (not password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Admin - list all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Admin - create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Admin - get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Admin - update user role or info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAdminUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── BORROWERS ───────────────────────────────────────────────────────────────
...
  /borrowers/{borrowerId}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List monthly payment records for a borrower
      parameters:
        - name: borrowerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a monthly payment
      parameters:
        - name: borrowerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /borrowers/{borrowerId}/payments/{paymentId}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment record
      parameters:
        - name: borrowerId
          in: path
          required: true
          schema:
            type: integer
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment record
      parameters:
        - name: borrowerId
          in: path
          required: true
          schema:
            type: integer
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ANALYTICS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, name, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, user]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    UserAdminUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean

    # ── Borrower ───────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, borrowerId, month, year, principalAmount, interestRate, baseInterestRate, commissionRate, interestAmount, baseInterestAmount, commissionAmount, isPaid, createdAt]
      properties:
        id:
          type: integer
        borrowerId:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        principalAmount:
          type: number
        interestRate:
          type: number
        baseInterestRate:
          type: number
        commissionRate:
          type: number
        interestAmount:
          type: number
          description: Total interest amount (principalAmount * interestRate / 100)
        baseInterestAmount:
          type: number
          description: Base interest portion (principalAmount * 10 / 100)
        commissionAmount:
          type: number
          description: Commission portion (principalAmount * commissionRate / 100)
        isPaid:
          type: boolean
        paidDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [month, year]
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        isPaid:
          type: boolean
        paidDate:
          type: string
          format: date
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        isPaid:
          type: boolean
        paidDate:
          type: string
          format: date
        notes:
          type: string

    # ── Analytics ──────────────────────────────────────────────────────────────
...