# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  borrowerId: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  principalAmount: number;
  interestRate: number;
  baseInterestRate: number;
  commissionRate: number;
  /** Total interest amount (principalAmount * interestRate / 100) */
  interestAmount: number;
  /** Base interest portion (principalAmount * 10 / 100) */
  baseInterestAmount: number;
  /** Commission portion (principalAmount * commissionRate / 100) */
  commissionAmount: number;
  isPaid: boolean;
  /** @nullable */
  paidDate?: Date | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}