# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  isActive?: boolean;
  createdAt: Date;
  updatedAt?: Date;
}