# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  phone?: string;
}